Moved validate_icon to its own file
